template_aiohttp_project/requirements.txt
aiohttp==3.6.2
aiohttp-session==2.9.0
cryptography==2.8
jinja2==2.11.1
aiohttp-jinja2==1.2.0
peewee==2.10.2
peewee-async==0.5.12
aiopg==1.0.0
click==7.0
python-dotenv==0.11.0
aiosqlite==0.11.0
ipython==7.11.1
paho-mqtt==1.5.0
aioredis==1.3.1